Changeover to NiceGUI interface
