--- 1.3.14
@@ -1,106 +1,0 @@
-Collecting nicegui
-  Downloading nicegui-2.17.0-py3-none-any.whl (13.5 MB)
-     --------------------------------------- 13.5/13.5 MB 11.3 MB/s eta 0:00:00
-Collecting starlette>=0.45.3
-  Using cached starlette-0.46.2-py3-none-any.whl (72 kB)
-Collecting wait_for2>=0.3.2
-  Using cached wait_for2-0.3.2.tar.gz (16 kB)
-  Preparing metadata (setup.py): started
-  Preparing metadata (setup.py): finished with status 'done'
-Collecting itsdangerous<3.0.0,>=2.1.2
-  Using cached itsdangerous-2.2.0-py3-none-any.whl (16 kB)
-Requirement already satisfied: requests>=2.32.0 in d:\repos\joycaptioner\venv\lib\site-packages (from nicegui) (2.32.3)
-Collecting orjson>=3.9.15
-  Using cached orjson-3.10.18-cp310-cp310-win_amd64.whl (134 kB)
-Collecting aiohttp>=3.10.2
-  Using cached aiohttp-3.11.18-cp310-cp310-win_amd64.whl (442 kB)
-Collecting httpx>=0.24.0
-  Using cached httpx-0.28.1-py3-none-any.whl (73 kB)
-Collecting python-socketio[asyncio-client]>=5.10.0
-  Using cached python_socketio-5.13.0-py3-none-any.whl (77 kB)
-Collecting watchfiles>=0.18.1
-  Using cached watchfiles-1.0.5-cp310-cp310-win_amd64.whl (291 kB)
-Requirement already satisfied: typing-extensions>=4.0.0 in d:\repos\joycaptioner\venv\lib\site-packages (from nicegui) (4.13.2)
-Collecting docutils>=0.19.0
-  Using cached docutils-0.21.2-py3-none-any.whl (587 kB)
-Collecting markdown2!=2.4.11,>=2.4.7
-  Using cached markdown2-2.5.3-py3-none-any.whl (48 kB)
-Collecting aiofiles>=23.1.0
-  Using cached aiofiles-24.1.0-py3-none-any.whl (15 kB)
-Collecting ifaddr>=0.2.0
-  Using cached ifaddr-0.2.0-py3-none-any.whl (12 kB)
-Collecting vbuild>=0.8.2
-  Using cached vbuild-0.8.2-py2.py3-none-any.whl (9.4 kB)
-Requirement already satisfied: urllib3!=2.0.0,!=2.0.1,!=2.0.2,!=2.0.3,!=2.0.4,!=2.0.5,!=2.0.6,!=2.0.7,!=2.1.0,!=2.2.0,!=2.2.1,>=1.26.18 in d:\repos\joycaptioner\venv\lib\site-packages (from nicegui) (2.4.0)
-Collecting uvicorn[standard]>=0.22.0
-  Using cached uvicorn-0.34.2-py3-none-any.whl (62 kB)
-Collecting h11>=0.16.0
-  Using cached h11-0.16.0-py3-none-any.whl (37 kB)
-Collecting fastapi>=0.109.1
-  Using cached fastapi-0.115.12-py3-none-any.whl (95 kB)
-Collecting python-multipart>=0.0.18
-  Using cached python_multipart-0.0.20-py3-none-any.whl (24 kB)
-Requirement already satisfied: certifi>=2024.07.04 in d:\repos\joycaptioner\venv\lib\site-packages (from nicegui) (2025.4.26)
-Collecting Pygments<3.0.0,>=2.15.1
-  Using cached pygments-2.19.1-py3-none-any.whl (1.2 MB)
-Collecting python-engineio>=4.12.0
-  Downloading python_engineio-4.12.1-py3-none-any.whl (59 kB)
-     ---------------------------------------- 59.3/59.3 kB ? eta 0:00:00
-Requirement already satisfied: jinja2<4.0.0,>=3.1.6 in d:\repos\joycaptioner\venv\lib\site-packages (from nicegui) (3.1.6)
-Collecting aiohappyeyeballs>=2.3.0
-  Using cached aiohappyeyeballs-2.6.1-py3-none-any.whl (15 kB)
-Collecting propcache>=0.2.0
-  Using cached propcache-0.3.1-cp310-cp310-win_amd64.whl (45 kB)
-Collecting frozenlist>=1.1.1
-  Using cached frozenlist-1.6.0-cp310-cp310-win_amd64.whl (120 kB)
-Collecting multidict<7.0,>=4.5
-  Downloading multidict-6.4.4-cp310-cp310-win_amd64.whl (38 kB)
-Collecting attrs>=17.3.0
-  Using cached attrs-25.3.0-py3-none-any.whl (63 kB)
-Collecting yarl<2.0,>=1.17.0
-  Using cached yarl-1.20.0-cp310-cp310-win_amd64.whl (92 kB)
-Collecting aiosignal>=1.1.2
-  Using cached aiosignal-1.3.2-py2.py3-none-any.whl (7.6 kB)
-Collecting async-timeout<6.0,>=4.0
-  Using cached async_timeout-5.0.1-py3-none-any.whl (6.2 kB)
-Collecting pydantic!=1.8,!=1.8.1,!=2.0.0,!=2.0.1,!=2.1.0,<3.0.0,>=1.7.4
-  Using cached pydantic-2.11.4-py3-none-any.whl (443 kB)
-Collecting httpcore==1.*
-  Using cached httpcore-1.0.9-py3-none-any.whl (78 kB)
-Collecting anyio
-  Using cached anyio-4.9.0-py3-none-any.whl (100 kB)
-Requirement already satisfied: idna in d:\repos\joycaptioner\venv\lib\site-packages (from httpx>=0.24.0->nicegui) (3.10)
-Requirement already satisfied: MarkupSafe>=2.0 in d:\repos\joycaptioner\venv\lib\site-packages (from jinja2<4.0.0,>=3.1.6->nicegui) (3.0.2)
-Collecting simple-websocket>=0.10.0
-  Using cached simple_websocket-1.1.0-py3-none-any.whl (13 kB)
-Collecting bidict>=0.21.0
-  Using cached bidict-0.23.1-py3-none-any.whl (32 kB)
-Requirement already satisfied: charset-normalizer<4,>=2 in d:\repos\joycaptioner\venv\lib\site-packages (from requests>=2.32.0->nicegui) (3.4.2)
-Collecting click>=7.0
-  Using cached click-8.2.0-py3-none-any.whl (102 kB)
-Collecting websockets>=10.4
-  Using cached websockets-15.0.1-cp310-cp310-win_amd64.whl (176 kB)
-Collecting httptools>=0.6.3
-  Using cached httptools-0.6.4-cp310-cp310-win_amd64.whl (88 kB)
-Requirement already satisfied: pyyaml>=5.1 in d:\repos\joycaptioner\venv\lib\site-packages (from uvicorn[standard]>=0.22.0->nicegui) (6.0.2)
-Requirement already satisfied: colorama>=0.4 in d:\repos\joycaptioner\venv\lib\site-packages (from uvicorn[standard]>=0.22.0->nicegui) (0.4.6)
-Collecting python-dotenv>=0.13
-  Downloading python_dotenv-1.1.0-py3-none-any.whl (20 kB)
-Collecting pscript<0.8.0,>=0.7.0
-  Using cached pscript-0.7.7-py3-none-any.whl (126 kB)
-Collecting exceptiongroup>=1.0.2
-  Using cached exceptiongroup-1.3.0-py3-none-any.whl (16 kB)
-Collecting sniffio>=1.1
-  Using cached sniffio-1.3.1-py3-none-any.whl (10 kB)
-Collecting typing-inspection>=0.4.0
-  Using cached typing_inspection-0.4.0-py3-none-any.whl (14 kB)
-Collecting pydantic-core==2.33.2
-  Using cached pydantic_core-2.33.2-cp310-cp310-win_amd64.whl (2.0 MB)
-Collecting annotated-types>=0.6.0
-  Using cached annotated_types-0.7.0-py3-none-any.whl (13 kB)
-Collecting wsproto
-  Using cached wsproto-1.2.0-py3-none-any.whl (24 kB)
-Installing collected packages: wait_for2, pscript, ifaddr, websockets, vbuild, typing-inspection, sniffio, python-multipart, python-dotenv, Pygments, pydantic-core, propcache, orjson, multidict, markdown2, itsdangerous, httptools, h11, frozenlist, exceptiongroup, docutils, click, bidict, attrs, async-timeout, annotated-types, aiohappyeyeballs, aiofiles, yarl, wsproto, uvicorn, pydantic, httpcore, anyio, aiosignal, watchfiles, starlette, simple-websocket, httpx, aiohttp, python-engineio, fastapi, python-socketio, nicegui
-  Running setup.py install for wait_for2: started
-  Running setup.py install for wait_for2: finished with status 'done'
-Successfully installed Pygments-2.19.1 aiofiles-24.1.0 aiohappyeyeballs-2.6.1 aiohttp-3.11.18 aiosignal-1.3.2 annotated-types-0.7.0 anyio-4.9.0 async-timeout-5.0.1 attrs-25.3.0 bidict-0.23.1 click-8.2.0 docutils-0.21.2 exceptiongroup-1.3.0 fastapi-0.115.12 frozenlist-1.6.0 h11-0.16.0 httpcore-1.0.9 httptools-0.6.4 httpx-0.28.1 ifaddr-0.2.0 itsdangerous-2.2.0 markdown2-2.5.3 multidict-6.4.4 nicegui-2.17.0 orjson-3.10.18 propcache-0.3.1 pscript-0.7.7 pydantic-2.11.4 pydantic-core-2.33.2 python-dotenv-1.1.0 python-engineio-4.12.1 python-multipart-0.0.20 python-socketio-5.13.0 simple-websocket-1.1.0 sniffio-1.3.1 starlette-0.46.2 typing-inspection-0.4.0 uvicorn-0.34.2 vbuild-0.8.2 wait_for2-0.3.2 watchfiles-1.0.5 websockets-15.0.1 wsproto-1.2.0 yarl-1.20.0
