+++ requirements.txt
@@ -12,7 +12,8 @@ torchvision
 torchaudio
 
 # Core dependencies
-PySide6
+nicegui>=1.3.14
+#PySide6
 Pillow>=9.0
 transformers>=4.36
